Add files via upload
POLYGENIC RISK SCORES AND MODELS

SNPs identified from the GWAS of the alcohol dependence and NESARC datasets were used to create PRS with the PRSice software (Euesden et al., 2014). Once these
scores were obtained, they were then entered in a logistic regression model, with sex and age as covariates. The model was trained with the PRS from the NESARC training dataset and was then used to predict alcohol dependence. The NESARC dataset, that contained PRS, sex, age and alcohol dependence data, was split into 80:20, for the training and test sets respectively. After the training of the model, it was then used to predict alcohol dependence in the NESARC test dataset. After alcohol dependence had been predicted, the test set was then used to see where the actual alcohol dependence and predicted alcohol dependence correlated.

From the model, it was predicted that alcohol dependence occurred 404 times within this dataset. When comparing predicted alcohol dependence to the initial value for
alcohol dependence, it was found that 199 records correlated. This model had the following values:
Accuracy: 66.80%
Sensitivity: 69.17%
Specificity: 49.26%

Once the model was trained on the NESARC dataset, it was then used to predict alcohol dependence in the alcohol dependence dataset. The model predicted 125
individuals to be alcohol dependent, however 86 were successful. The model performed according to the following statistics:
Accuracy: 25.84%
Sensitivity: 22.83%
Specificity: 68.80%.

From these 125 individuals it was then determined where alcohol and gambling dependence coincided, meaning that they occurred together in an individual. There
were 9 individuals that were successfully predicted to have alcohol dependence and a gambling addiction. This is compared to the dataset actually having 188 individuals who were addicted to AUD and GD.