Add files via upload
PRS of Alcohol Dependence dataset in non-African Americans and African Americans

For their respective datasets for non-African Americans and African Americans, the PRS were generated with the PRSice software (Euesden, Lewis and O’Reilly, 2014).
These scores were then entered into a logistic regression model with sex and age used as covariates. Again, the data was split into a training and test set, using 80:20,
allowing the model to be trained on the training data and tested on the test data, where alcohol was used as the target variable. The statistics for the models are given below.

Non-African American population model
It was predicted that alcohol dependence occurred 204 times from the training set. When comparing predicted alcohol dependence to the actual value for alcohol
dependence it was found that 204 records correlated. The model performed according to the following statistics:
Accuracy: 99.03%
Sensitivity: 33.333%
Specificity: 100%