Add files via upload
Applying population models

The model that was trained on the African American dataset was used to predict alcohol dependence in the non-African American dataset, this shall be referred to as
Model 1. The reverse was also applied and shall be referred to as Model 2. 

The model statistics were calculated: 

In the non-African American population, there were 84 individuals who suffered from
both AUD and GD. All 84 individuals were predicted successfully to have AUD and
GD. In the African American population, there were 187 individuals who suffered from
both AUD and GD. Again, all 187 individual were predicted successfully to have AUD
and GD based on the model.